+++ lib/core/alice_dio_interceptor.dart
@@ -20,82 +20,97 @@ class AliceDioInterceptor extends InterceptorsWrapper {
 
   @override
   void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
-    final AliceHttpCall call = AliceHttpCall(options.hashCode);
+    final call = AliceHttpCall(options.hashCode);
 
-    final Uri uri = options.uri;
+    final uri = options.uri;
     call.method = options.method;
     var path = options.uri.path;
     if (path.isEmpty) {
       path = '/';
     }
-    call.endpoint = path;
-    call.server = uri.host;
-    call.client = 'Dio';
-    call.uri = options.uri.toString();
+    call
+      ..endpoint = path
+      ..server = uri.host
+      ..client = 'Dio'
+      ..uri = options.uri.toString();
 
     if (uri.scheme == 'https') {
       call.secure = true;
     }
 
-    final AliceHttpRequest request = AliceHttpRequest();
+    final request = AliceHttpRequest();
 
     final data = options.data;
     if (data == null) {
-      request.size = 0;
-      request.body = '';
+      request
+        ..size = 0
+        ..body = '';
     } else {
       if (data is FormData) {
         request.body += 'Form data';
 
         if (data.fields.isNotEmpty == true) {
-          final List<AliceFormDataField> fields = [];
-          for (var entry in data.fields) {
+          final fields = <AliceFormDataField>[];
+          for (final entry in data.fields) {
             fields.add(AliceFormDataField(entry.key, entry.value));
           }
           request.formDataFields = fields;
         }
         if (data.files.isNotEmpty == true) {
-          final List<AliceFormDataFile> files = [];
-          for (var entry in data.files) {
-            files.add(AliceFormDataFile(entry.value.filename ?? '',
-                entry.value.contentType.toString(), entry.value.length));
+          final files = <AliceFormDataFile>[];
+          for (final entry in data.files) {
+            files.add(
+              AliceFormDataFile(
+                entry.value.filename ?? '',
+                entry.value.contentType.toString(),
+                entry.value.length,
+              ),
+            );
           }
 
           request.formDataFiles = files;
         }
       } else {
-        request.size = utf8.encode(data.toString()).length;
-        request.body = data;
+        request
+          ..size = utf8.encode(data.toString()).length
+          ..body = data;
       }
     }
 
-    request.time = DateTime.now();
-    request.headers = options.headers;
-    request.contentType = options.contentType.toString();
-    request.queryParameters = options.queryParameters;
+    request
+      ..time = DateTime.now()
+      ..headers = options.headers
+      ..contentType = options.contentType.toString()
+      ..queryParameters = options.queryParameters;
 
-    call.request = request;
-    call.response = AliceHttpResponse();
+    call
+      ..request = request
+      ..response = AliceHttpResponse();
 
     aliceCore.addCall(call);
     super.onRequest(options, handler);
   }
 
   @override
-  void onResponse(Response response, ResponseInterceptorHandler handler) {
-    final httpResponse = AliceHttpResponse();
-    httpResponse.status = response.statusCode ?? -1;
+  void onResponse(
+    Response<dynamic> response,
+    ResponseInterceptorHandler handler,
+  ) {
+    final httpResponse = AliceHttpResponse()
+      ..status = response.statusCode ?? -1;
 
     if (response.data == null) {
-      httpResponse.body = '';
-      httpResponse.size = 0;
+      httpResponse
+        ..body = ''
+        ..size = 0;
     } else {
-      httpResponse.body = response.data;
-      httpResponse.size = utf8.encode(response.data.toString()).length;
+      httpResponse
+        ..body = response.data
+        ..size = utf8.encode(response.data.toString()).length;
     }
 
     httpResponse.time = DateTime.now();
-    final Map<String, String> headers = {};
+    final headers = <String, String>{};
     response.headers.forEach((header, values) {
       headers[header] = values.toString();
     });
@@ -109,16 +124,14 @@ class AliceDioInterceptor extends InterceptorsWrapper {
 
   @override
   void onError(DioException error, ErrorInterceptorHandler handler) {
-    final httpError = AliceHttpError();
-    httpError.error = error.toString();
+    final httpError = AliceHttpError()..error = error.toString();
     if (error is Error) {
       final basicError = error as Error;
       httpError.stackTrace = basicError.stackTrace;
     }
 
     aliceCore.addError(httpError, error.requestOptions.hashCode);
-    final httpResponse = AliceHttpResponse();
-    httpResponse.time = DateTime.now();
+    final httpResponse = AliceHttpResponse()..time = DateTime.now();
     if (error.response == null) {
       httpResponse.status = -1;
       aliceCore.addResponse(httpResponse, error.requestOptions.hashCode);
@@ -126,20 +139,23 @@ class AliceDioInterceptor extends InterceptorsWrapper {
       httpResponse.status = error.response?.statusCode ?? -1;
 
       if (error.response?.data == null) {
-        httpResponse.body = '';
-        httpResponse.size = 0;
+        httpResponse
+          ..body = ''
+          ..size = 0;
       } else {
-        httpResponse.body = error.response?.data;
-        httpResponse.size =
-            utf8.encode(error.response?.data.toString() ?? '').length;
+        httpResponse
+          ..body = error.response?.data
+          ..size = utf8.encode(error.response?.data.toString() ?? '').length;
       }
-      final Map<String, String> headers = {};
+      final headers = <String, String>{};
       error.response?.headers.forEach((header, values) {
         headers[header] = values.toString();
       });
       httpResponse.headers = headers;
       aliceCore.addResponse(
-          httpResponse, error.response?.requestOptions.hashCode);
+        httpResponse,
+        error.response?.requestOptions.hashCode,
+      );
     }
     super.onError(error, handler);
   }