+++ lib/helper/alice_save_helper.dart
@@ -7,69 +7,88 @@ mixin AliceSaveHelper {
   static const JsonEncoder _encoder = JsonEncoder.withIndent('  ');
 
   static Future<String> _buildAliceLog() async {
-    final StringBuffer stringBuffer = StringBuffer();
-    stringBuffer.write('Alice - HTTP Inspector\n');
-    stringBuffer.write('Generated: ${DateTime.now().toIso8601String()}\n');
-    stringBuffer.write('\n');
+    final stringBuffer = StringBuffer()
+      ..write('Alice - HTTP Inspector\n')
+      ..write('Generated: ${DateTime.now().toIso8601String()}\n')
+      ..write('\n');
     return stringBuffer.toString();
   }
 
   static String _buildCallLog(AliceHttpCall call) {
-    final StringBuffer stringBuffer = StringBuffer();
-    stringBuffer.write('===========================================\n');
-    stringBuffer.write('Id: ${call.id}\n');
-    stringBuffer.write('============================================\n');
-    stringBuffer.write('--------------------------------------------\n');
-    stringBuffer.write('General data\n');
-    stringBuffer.write('--------------------------------------------\n');
-    stringBuffer.write('Server: ${call.server} \n');
-    stringBuffer.write('Method: ${call.method} \n');
-    stringBuffer.write('Endpoint: ${call.endpoint} \n');
-    stringBuffer.write('Client: ${call.client} \n');
+    final stringBuffer = StringBuffer()
+      ..write('===========================================\n')
+      ..write('Id: ${call.id}\n')
+      ..write('============================================\n')
+      ..write('--------------------------------------------\n')
+      ..write('General data\n')
+      ..write('--------------------------------------------\n')
+      ..write('Server: ${call.server} \n')
+      ..write('Method: ${call.method} \n')
+      ..write('Endpoint: ${call.endpoint} \n')
+      ..write('Client: ${call.client} \n')
+      ..write('Duration ${AliceConversionHelper.formatTime(call.duration)}\n')
+      ..write('Secured connection: ${call.secure}\n')
+      ..write('Completed: ${!call.loading} \n')
+      ..write('--------------------------------------------\n')
+      ..write('Request\n')
+      ..write('--------------------------------------------\n')
+      ..write('Request time: ${call.request!.time}\n')
+      ..write('Request content type: ${call.request!.contentType}\n')
+      ..write('Request cookies: ${_encoder.convert(call.request!.cookies)}\n')
+      ..write('Request headers: ${_encoder.convert(call.request!.headers)}\n');
+    if (call.request!.queryParameters.isNotEmpty) {
+      stringBuffer.write(
+        'Request query params: '
+        '${_encoder.convert(call.request!.queryParameters)}\n',
+      );
+    }
     stringBuffer
-        .write('Duration ${AliceConversionHelper.formatTime(call.duration)}\n');
-    stringBuffer.write('Secured connection: ${call.secure}\n');
-    stringBuffer.write('Completed: ${!call.loading} \n');
-    stringBuffer.write('--------------------------------------------\n');
-    stringBuffer.write('Request\n');
-    stringBuffer.write('--------------------------------------------\n');
-    stringBuffer.write('Request time: ${call.request?.time}\n');
-    stringBuffer.write('Request content type: ${call.request?.contentType}\n');
-    stringBuffer
-        .write('Request cookies: ${_encoder.convert(call.request?.cookies)}\n');
-    stringBuffer
-        .write('Request headers: ${_encoder.convert(call.request?.headers)}\n');
-    stringBuffer.write(
-        'Request size: ${AliceConversionHelper.formatBytes(call.request?.size ?? 0)}\n');
-    stringBuffer.write(
-        'Request body: ${AliceParser.formatBody(call.request?.body, AliceParser.getContentType(call.request?.headers ?? {}))}\n');
-    stringBuffer.write('--------------------------------------------\n');
-    stringBuffer.write('Response\n');
-    stringBuffer.write('--------------------------------------------\n');
-    stringBuffer.write('Response time: ${call.response?.time}\n');
-    stringBuffer.write('Response status: ${call.response?.status}\n');
-    stringBuffer.write(
-        'Response size: ${AliceConversionHelper.formatBytes(call.response?.size ?? 0)}\n');
-    stringBuffer.write(
-        'Response headers: ${_encoder.convert(call.response?.headers)}\n');
-    stringBuffer.write(
-        'Response body: ${AliceParser.formatBody(call.response?.body, AliceParser.getContentType(call.response?.headers ?? {}))}\n');
+      ..write(
+        'Request size: '
+        '${AliceConversionHelper.formatBytes(call.request!.size)}\n',
+      )
+      ..write(
+        'Request body: ${AliceParser.formatBody(
+          call.request!.body,
+          AliceParser.getContentType(call.request!.headers),
+        )}\n',
+      )
+      ..write('--------------------------------------------\n')
+      ..write('Response\n')
+      ..write('--------------------------------------------\n')
+      ..write('Response time: ${call.response!.time}\n')
+      ..write('Response status: ${call.response!.status}\n')
+      ..write(
+        'Response size: '
+        '${AliceConversionHelper.formatBytes(call.response!.size)}\n',
+      )
+      ..write(
+        'Response headers: ${_encoder.convert(call.response!.headers)}\n',
+      )
+      ..write(
+        'Response body: ${AliceParser.formatBody(
+          call.response!.body,
+          AliceParser.getContentType(call.response!.headers),
+        )}\n',
+      );
     if (call.error != null) {
-      stringBuffer.write('--------------------------------------------\n');
-      stringBuffer.write('Error\n');
-      stringBuffer.write('--------------------------------------------\n');
-      stringBuffer.write('Error: ${call.error?.error}\n');
-      if (call.error?.stackTrace != null) {
-        stringBuffer.write('Error stacktrace: ${call.error?.stackTrace}\n');
+      stringBuffer
+        ..write('--------------------------------------------\n')
+        ..write('Error\n')
+        ..write('--------------------------------------------\n')
+        ..write('Error: ${call.error!.error}\n');
+      if (call.error!.stackTrace != null) {
+        stringBuffer.write('Error stacktrace: ${call.error!.stackTrace}\n');
       }
     }
-    stringBuffer.write('--------------------------------------------\n');
-    stringBuffer.write('Curl\n');
-    stringBuffer.write('--------------------------------------------\n');
-    stringBuffer.write(call.getCurlCommand());
-    stringBuffer.write('\n');
-    stringBuffer.write('==============================================\n');
-    stringBuffer.write('\n');
+    stringBuffer
+      ..write('--------------------------------------------\n')
+      ..write('Curl\n')
+      ..write('--------------------------------------------\n')
+      ..write(call.getCurlCommand())
+      ..write('\n')
+      ..write('==============================================\n')
+      ..write('\n');
 
     return stringBuffer.toString();
   }