+++ lib/ui/page/alice_call_details_screen.dart
@@ -3,11 +3,11 @@
 import 'package:alice/core/alice_core.dart';
 import 'package:alice/helper/alice_save_helper.dart';
 import 'package:alice/model/alice_http_call.dart';
-import 'package:alice/utils/alice_constants.dart';
 import 'package:alice/ui/widget/alice_call_error_widget.dart';
 import 'package:alice/ui/widget/alice_call_overview_widget.dart';
 import 'package:alice/ui/widget/alice_call_request_widget.dart';
 import 'package:alice/ui/widget/alice_call_response_widget.dart';
+import 'package:alice/utils/alice_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:share_plus/share_plus.dart';
 
@@ -41,22 +41,19 @@ class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
               secondary: AliceConstants.lightRed,
             ),
         tabBarTheme: const TabBarTheme(
-            indicator: UnderlineTabIndicator(
-              borderSide: BorderSide(
-                width: 2,
-                color: Colors.white,
-              ),
-            ),
-            indicatorSize: TabBarIndicatorSize.tab),
+          indicator: UnderlineTabIndicator(),
+          indicatorSize: TabBarIndicatorSize.tab,
+        ),
       ),
       child: StreamBuilder<List<AliceHttpCall>>(
         stream: widget.core.callsSubject,
         initialData: [widget.call],
         builder: (context, callsSnapshot) {
           if (callsSnapshot.hasData) {
-            final AliceHttpCall? call = callsSnapshot.data?.firstWhere(
-                (snapshotCall) => snapshotCall.id == widget.call.id,
-                orElse: null);
+            final call = callsSnapshot.data?.firstWhere(
+              (snapshotCall) => snapshotCall.id == widget.call.id,
+              orElse: null,
+            );
             if (call != null) {
               return _buildMainWidget();
             } else {
@@ -78,8 +75,10 @@ class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
           backgroundColor: AliceConstants.lightRed,
           key: const Key('share_key'),
           onPressed: () async {
-            Share.share(await _getSharableResponseString(),
-                subject: 'Request Details');
+            await Share.share(
+              await _getSharableResponseString(),
+              subject: 'Request Details',
+            );
           },
           child: const Icon(Icons.share),
         ),
@@ -106,25 +105,23 @@ class _AliceCallDetailsScreenState extends State<AliceCallDetailsScreen>
   }
 
   List<Widget> _getTabBars() {
-    final List<Widget> widgets = [];
-    widgets.add(const Tab(icon: Icon(Icons.info_outline), text: 'Overview'));
-    widgets.add(const Tab(icon: Icon(Icons.arrow_upward), text: 'Request'));
-    widgets.add(const Tab(icon: Icon(Icons.arrow_downward), text: 'Response'));
-    widgets.add(
+    return <Widget>[
+      const Tab(icon: Icon(Icons.info_outline), text: 'Overview'),
+      const Tab(icon: Icon(Icons.arrow_upward), text: 'Request'),
+      const Tab(icon: Icon(Icons.arrow_downward), text: 'Response'),
       const Tab(
         icon: Icon(Icons.warning),
         text: 'Error',
       ),
-    );
-    return widgets;
+    ];
   }
 
   List<Widget> _getTabBarViewList() {
-    final List<Widget> widgets = [];
-    widgets.add(AliceCallOverviewWidget(widget.call));
-    widgets.add(AliceCallRequestWidget(widget.call));
-    widgets.add(AliceCallResponseWidget(widget.call));
-    widgets.add(AliceCallErrorWidget(widget.call));
-    return widgets;
+    return [
+      AliceCallOverviewWidget(widget.call),
+      AliceCallRequestWidget(widget.call),
+      AliceCallResponseWidget(widget.call),
+      AliceCallErrorWidget(widget.call),
+    ];
   }
 }