+++ lib/ui/page/alice_calls_list_screen.dart
@@ -1,13 +1,12 @@
-import 'package:alice/model/alice_menu_item.dart';
-import 'package:alice/helper/alice_alert_helper.dart';
-import 'package:alice/ui/page/alice_call_details_screen.dart';
 import 'package:alice/core/alice_core.dart';
+import 'package:alice/helper/alice_alert_helper.dart';
 import 'package:alice/model/alice_http_call.dart';
-import 'package:alice/utils/alice_constants.dart';
+import 'package:alice/model/alice_menu_item.dart';
+import 'package:alice/ui/page/alice_call_details_screen.dart';
+import 'package:alice/ui/page/alice_stats_screen.dart';
 import 'package:alice/ui/widget/alice_call_list_item_widget.dart';
+import 'package:alice/utils/alice_constants.dart';
 import 'package:flutter/material.dart';
-
-import 'alice_stats_screen.dart';
 
 class AliceCallsListScreen extends StatefulWidget {
   final AliceCore _aliceCore;
@@ -26,8 +25,9 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
   final List<AliceMenuItem> _menuItems = [];
 
   _AliceCallsListScreenState() {
-    _menuItems.add(AliceMenuItem('Delete', Icons.delete));
-    _menuItems.add(AliceMenuItem('Stats', Icons.insert_chart));
+    _menuItems
+      ..add(AliceMenuItem('Delete', Icons.delete))
+      ..add(AliceMenuItem('Stats', Icons.insert_chart));
   }
 
   @override
@@ -37,9 +37,10 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
         useMaterial3: false,
         brightness: widget._aliceCore.brightness,
         colorScheme: Theme.of(context).colorScheme.copyWith(
-            brightness: widget._aliceCore.brightness,
-            primary: AliceConstants.strongRed,
-            secondary: AliceConstants.lightRed),
+              brightness: widget._aliceCore.brightness,
+              primary: AliceConstants.strongRed,
+              secondary: AliceConstants.lightRed,
+            ),
       ),
       child: Scaffold(
         appBar: AppBar(
@@ -78,21 +79,23 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
 
   Widget _buildMenuButton() {
     return PopupMenuButton<AliceMenuItem>(
-      onSelected: (AliceMenuItem item) => _onMenuItemSelected(item),
+      onSelected: _onMenuItemSelected,
       itemBuilder: (BuildContext context) {
         return _menuItems.map((AliceMenuItem item) {
           return PopupMenuItem<AliceMenuItem>(
             value: item,
-            child: Row(children: [
-              Icon(
-                item.iconData,
-                color: AliceConstants.lightRed,
-              ),
-              const Padding(
-                padding: EdgeInsets.only(left: 10),
-              ),
-              Text(item.title)
-            ]),
+            child: Row(
+              children: [
+                Icon(
+                  item.iconData,
+                  color: AliceConstants.lightRed,
+                ),
+                const Padding(
+                  padding: EdgeInsets.only(left: 10),
+                ),
+                Text(item.title),
+              ],
+            ),
           );
         }).toList();
       },
@@ -109,10 +112,10 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
       autofocus: true,
       decoration: InputDecoration(
         hintText: 'Search http request...',
-        hintStyle: TextStyle(fontSize: 16.0, color: AliceConstants.grey),
+        hintStyle: TextStyle(fontSize: 16, color: AliceConstants.grey),
         border: InputBorder.none,
       ),
-      style: TextStyle(fontSize: 16.0, color: AliceConstants.white),
+      style: TextStyle(fontSize: 16, color: AliceConstants.white),
       cursorColor: AliceConstants.white,
       onChanged: _updateSearchQuery,
     );
@@ -131,12 +134,14 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
     return StreamBuilder<List<AliceHttpCall>>(
       stream: aliceCore.callsSubject,
       builder: (context, snapshot) {
-        List<AliceHttpCall> calls = snapshot.data ?? [];
-        final String query = _queryTextEditingController.text.trim();
+        var calls = snapshot.data ?? [];
+        final query = _queryTextEditingController.text.trim();
         if (query.isNotEmpty) {
           calls = calls
-              .where((call) =>
-                  call.endpoint.toLowerCase().contains(query.toLowerCase()))
+              .where(
+                (call) =>
+                    call.endpoint.toLowerCase().contains(query.toLowerCase()),
+              )
               .toList();
         }
         if (calls.isNotEmpty) {
@@ -152,35 +157,41 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
     return Container(
       margin: const EdgeInsets.symmetric(horizontal: 32),
       child: Center(
-        child: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
-          Icon(
-            Icons.error_outline,
-            color: AliceConstants.orange,
-          ),
-          const SizedBox(height: 6),
-          const Text(
-            'There are no calls to show',
-            style: TextStyle(fontSize: 18),
-          ),
-          const SizedBox(height: 12),
-          const Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-            Text(
-              '• Check if you send any http request',
-              style: TextStyle(fontSize: 12),
-              textAlign: TextAlign.center,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(
+              Icons.error_outline,
+              color: AliceConstants.orange,
             ),
-            Text(
-              '• Check your Alice configuration',
-              style: TextStyle(fontSize: 12),
-              textAlign: TextAlign.center,
+            const SizedBox(height: 6),
+            const Text(
+              'There are no calls to show',
+              style: TextStyle(fontSize: 18),
             ),
-            Text(
-              '• Check search filters',
-              style: TextStyle(fontSize: 12),
-              textAlign: TextAlign.center,
-            )
-          ])
-        ]),
+            const SizedBox(height: 12),
+            const Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  '• Check if you send any http request',
+                  style: TextStyle(fontSize: 12),
+                  textAlign: TextAlign.center,
+                ),
+                Text(
+                  '• Check your Alice configuration',
+                  style: TextStyle(fontSize: 12),
+                  textAlign: TextAlign.center,
+                ),
+                Text(
+                  '• Check search filters',
+                  style: TextStyle(fontSize: 12),
+                  textAlign: TextAlign.center,
+                ),
+              ],
+            ),
+          ],
+        ),
       ),
     );
   }
@@ -195,7 +206,7 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
   }
 
   void _onListItemClicked(AliceHttpCall call) {
-    Navigator.push(
+    Navigator.push<void>(
       widget._aliceCore.getContext()!,
       MaterialPageRoute(
         builder: (context) => AliceCallDetailsScreen(call, widget._aliceCore),
@@ -209,9 +220,9 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
       'Delete calls',
       'Do you want to delete http calls?',
       firstButtonTitle: 'No',
-      firstButtonAction: () => {},
+      firstButtonAction: () => <String, dynamic>{},
       secondButtonTitle: 'Yes',
-      secondButtonAction: () => _removeCalls(),
+      secondButtonAction: _removeCalls,
     );
   }
 
@@ -220,7 +231,7 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
   }
 
   void _showStatsScreen() {
-    Navigator.push(
+    Navigator.push<void>(
       aliceCore.getContext()!,
       MaterialPageRoute(
         builder: (context) => AliceStatsScreen(widget._aliceCore),