+++ lib/ui/widget/alice_call_request_widget.dart
@@ -19,21 +19,21 @@ class _AliceCallRequestWidget
 
   @override
   Widget build(BuildContext context) {
-    List<Widget> rows = [];
-    rows.add(getListRow('Started:', _call.request?.time.toString() ?? ''));
-    rows.add(getListRow('Bytes sent:', formatBytes(_call.request?.size ?? 0)));
-    rows.add(getListRow(
-        'Content type:', getContentType(_call.request?.headers ?? {})));
+    final rows = [
+      getListRow('Started:', _call.request!.time.toString()),
+      getListRow('Bytes sent:', formatBytes(_call.request!.size)),
+      getListRow('Content type:', getContentType(_call.request!.headers)),
+    ];
 
-    var body = _call.request?.body;
+    final body = _call.request?.body;
     var bodyContent = 'Body is empty';
     if (body != null) {
       bodyContent =
           formatBody(body, getContentType(_call.request?.headers ?? {}));
     }
     rows.add(getListRow('Body:', bodyContent));
-    var formDataFields = _call.request?.formDataFields;
-    if (formDataFields?.isNotEmpty == true) {
+    final formDataFields = _call.request?.formDataFields;
+    if (formDataFields?.isNotEmpty ?? false) {
       rows.add(getListRow('Form data fields: ', ''));
       formDataFields?.forEach(
         (field) {
@@ -41,18 +41,22 @@ class _AliceCallRequestWidget
         },
       );
     }
-    var formDataFiles = _call.request?.formDataFiles;
-    if (formDataFiles?.isNotEmpty == true) {
+    final formDataFiles = _call.request?.formDataFiles;
+    if (formDataFiles?.isNotEmpty ?? false) {
       rows.add(getListRow('Form data files: ', ''));
       formDataFiles?.forEach(
         (field) {
-          rows.add(getListRow('   • ${field.fileName}:',
-              '${field.contentType} / ${field.length} B'));
+          rows.add(
+            getListRow(
+              '   • ${field.fileName}:',
+              '${field.contentType} / ${field.length} B',
+            ),
+          );
         },
       );
     }
 
-    var headers = _call.request?.headers;
+    final headers = _call.request?.headers;
     var headersContent = 'Headers are empty';
     if (headers != null && headers.isNotEmpty) {
       headersContent = '';
@@ -61,7 +65,7 @@ class _AliceCallRequestWidget
     _call.request?.headers.forEach((header, value) {
       rows.add(getListRow('   • $header:', value.toString()));
     });
-    var queryParameters = _call.request?.queryParameters;
+    final queryParameters = _call.request?.queryParameters;
     var queryParametersContent = 'Query parameters are empty';
     if (queryParameters != null && queryParameters.isNotEmpty) {
       queryParametersContent = '';