+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Fix search field cursor color
+* Add very_good_analysis as linter rules
+
 ## 0.0.1
 
 * Initial release.